refactor: extract Ollama HTTP client
